clean up ExampleScreen1 + ExampleScreen2
