+++ example/lib/example_screen_1.dart
@@ -1,8 +1,8 @@
-import 'package:auto_size_text/auto_size_text.dart';
 import 'package:example/example_data_model.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_menu_morph/flutter_menu_morph.dart';
-import 'package:svg_flutter/svg_flutter.dart';
+
+import 'car_category_mixin.dart';
 
 /// Fetch data from BE
 class ExampleScreen1 extends StatefulWidget {
@@ -18,7 +18,7 @@ class ExampleScreen1 extends StatefulWidget {
   State<ExampleScreen1> createState() => _ExampleScreen1State();
 }
 
-class _ExampleScreen1State extends State<ExampleScreen1> {
+class _ExampleScreen1State extends State<ExampleScreen1> with CarCategoryMixin {
   late MenuBox2DController<CarCategory> _menuController;
 
   void _onMenuCreated(
@@ -28,29 +28,8 @@ class _ExampleScreen1State extends State<ExampleScreen1> {
 
     // fetch data
     Future.delayed(const Duration(seconds: 5)).then((_) {
-      _menuController.updateMenuBoardData(
-        MenuBoardData(
-          parent: _getMenuItem(CarCategory.parent()),
-          children: [
-            _getMenuItem(CarCategory.child1()),
-            _getMenuItem(CarCategory.child2()),
-            _getMenuItem(CarCategory.child3()),
-            _getMenuItem(CarCategory.child4()),
-            _getMenuItem(CarCategory.child5()),
-            _getMenuItem(CarCategory.child6()),
-          ],
-        ),
-      );
+      _menuController.updateMenuBoardData(getExampleData());
     });
-  }
-
-  MenuItem<CarCategory> _getMenuItem(CarCategory category) {
-    return MenuItem(
-      data: category,
-      itemBuilder: (_, category) {
-        return _buildCategory(category);
-      },
-    );
   }
 
   @override
@@ -70,27 +49,6 @@ class _ExampleScreen1State extends State<ExampleScreen1> {
           );
         },
       ),
-    );
-  }
-
-  Widget _buildCategory(CarCategory category) {
-    return Column(
-      mainAxisAlignment: MainAxisAlignment.center,
-      children: [
-        AutoSizeText(
-          category.name,
-          style: const TextStyle(
-            fontSize: 16.0,
-            color: Colors.black,
-          ),
-        ),
-        const SizedBox(height: 8.0),
-        SvgPicture.asset(
-          category.imagePath,
-          width: 32.0,
-          height: 32.0,
-        ),
-      ],
     );
   }
 }
